#9 and #10
Split namespace into glue and glueExtras, moved CRUDder
Got fancy routing tools out of glue\Route and put them in
glueExtras\RouteTools
