+++ _glue-site/config/glueExtras.yaml
@@ -4,6 +4,9 @@
 # This file is not licensed along with the rest of the Glue Framework. It is
 # released into the public domain, and you are free to modify or re-use it
 # in any way.
-Route:
-  routes:
-    - example
+glueExtras:
+  RouteTools:
+    staticExtensions:
+      # Add your own static extensions by adding them and their MIME as shown
+      # To disable a default extension, set its MIME to "false"
+      html: text/html
