+++ _glue/classes/glue/Route.php
@@ -20,7 +20,6 @@
 namespace glue;
 
 use \glue\Template;
-use \Mni\FrontYAML\Parser;
 use \Nanite;
 
 class Route
@@ -35,8 +34,8 @@ class Route
     }
     public static function any($url, $handler)
     {
-        Nanite::get($url, $handler);
-        Nanite::post($url, $handler);
+        static::get($url, $handler);
+        static::post($url, $handler);
     }
     public static function loadfile($file)
     {
@@ -46,86 +45,9 @@ class Route
     {
         return Nanite::$routeProccessed;
     }
-    public static function routeRedirects()
-    {
-        if ($filename = static::findFileForRoute(static::requestUri(), Conf::get('Route/content/path'), array('url'))) {
-            $file = fopen($filename, 'r');
-            $firstline = trim(fgets($file));
-            fclose($file);
-            header('Location: ' . $firstline);
-            Template::rawOutput();
-        }
-    }
-    public static function routeMarkdown()
-    {
-        if ($filename = static::findFileForRoute(static::requestUri(), Conf::get('Route/content/path'), array('md'))) {
-            $parser = new Parser();
-            if (!($document = $parser->parse(file_get_contents($filename)))) {
-                echo "<div><strong>Error:</strong> Failed to parse markdown/front-YAML</div>";
-                return;
-            }
-            static::any(static::requestUri(), function () use ($document) {
-                Template::setMulti($document->getYAML());
-                echo $document->getContent();
-            });
-        }
-    }
-    public static function routeStatic()
-    {
-        //This function should die as soon as it handles a file, because that way you skip templating
-        $filename = Conf::get('Route/content/path') . static::requestUri();
-        $extension = explode('.', $filename);
-        $extension = array_pop($extension);
-        if (array_key_exists($extension, Conf::get('Route/staticExtensions'))) {
-            if (Conf::get('Route/staticExtensions')[$extension] && is_file($filename)) {
-                header("Content-Type: " . Conf::get('Route/staticExtensions')[$extension]);
-                Template::rawOutput(file_get_contents($filename));
-            }
-        }
-    }
-    public static function routeCodepages()
-    {
-        $filename = static::findFileForRoute(static::requestUri(), Conf::get('Route/codepages/path'), array('php'));
-        if ($filename) {
-            static::any(static::requestUri(), function() use ($filename) {
-                require($filename);
-            });
-        }
-    }
-    public static function routeAutoRoute()
-    {
-        $class = explode('/', static::requestUri());
-        $class = strtolower($class[1]);
-        $class = preg_replace('/[^a-z0-9_\-]/i', '', $class);
-        $class = preg_replace('/\-+/', ' ', $class);
-        $class = ucwords($class);
-        $class = preg_replace('/ +/', '', $class);
-        $class = '\\AutoRoute\\' . $class;
-        if (class_exists($class, true)) {
-            if (method_exists($class, 'main')) {
-                static::any(static::requestUri().'(/.*)?', $class . '::main');
-            }
-        }
-    }
     public static function requestUri()
     {
         return Nanite::requestUri();
-    }
-    protected static function findFileForRoute($uri, $dir, $extensions)
-    {
-        if ($uri == '/') {
-            $uri = '';
-        }
-        foreach ($extensions as $ext) {
-            $possibilities = array();
-            $possibilities[] = $dir . $uri . '.' . $ext;
-            $possibilities[] = $dir . $uri . '/index.' . $ext;
-            foreach ($possibilities as $file) {
-                if (file_exists($file)) {
-                    return $file;
-                }
-            }
-        }
     }
 }
