+++ _glue/classes/glueExtras/CRUDder/CRUDder.php
@@ -17,7 +17,7 @@
   * with this program; if not, write to the Free Software Foundation, Inc.,
   * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-namespace glue\CRUDder;
+namespace glueExtras\CRUDder;
 
 use \Symfony\Component\Yaml\Yaml;
