+++ _glue/classes/glueExtras/CRUDder/DB.php
@@ -17,7 +17,8 @@
   * with this program; if not, write to the Free Software Foundation, Inc.,
   * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-namespace glue\CRUDder;
+namespace glueExtras\CRUDder;
+
 class DB
 {
     private static $conns = array();
