+++ phpunit.xml
@@ -1,8 +1,8 @@
 <phpunit bootstrap="tests/bootstrap.php">
   <testsuites>
     <testsuite name="CRUDder">
-      <file>tests/glue/CRUDder/CRUDderTest.php</file>
-      <file>tests/glue/CRUDder/CRUDderTransactionsTest.php</file>
+      <file>tests/glueExtras/CRUDder/CRUDderTest.php</file>
+      <file>tests/glueExtras/CRUDder/CRUDderTransactionsTest.php</file>
     </testsuite>
   </testsuites>
 </phpunit>
